Add api-builder to generate the API interfaces
The custom types generated won't actually work just yet since we need
to add some fun to the proxy to shim in `this` as the first argument
(plus serialize and deserialize the types) but we'll get there!

This is a fun little trick where we write the minimum API necessary
to generate the full API. We use the API to generate the API—it's
API-ception!