Attempt to support integration tests
Right now we're in a place where we *should* be able to run integration
tests against an embedded neovim (modulo filling the buffer, of course)
but the `nvim_buf_attach` call never seems to result in a Response
packet. Included in this commit is a unit test demonstrating the issue.

See neovim/neovim#8634