Hackily handle missing nvim_buf_* responses; hackily fix int. tests!
The test passes! Woah!

I keep going back and forth but while it's nice to be able to test vim
keystrokes I'm leaning towards this integration testing being a bad
approach for general testing. The mechanism we're using to dispatch the
received events is just for POC and is *terrible*....