Refactor API setup + events to be pluggable; make int. tests sane
By plugging in a different EventDaemon, in particular, we can save
ourselves from the headache of actually having a daemon coroutine that
wants to run on the UI thread and that we don't know when events might
arrive. Instead, once we've performed the actions we want to perform we
just process *all* the events until they stop coming. This actually
feels pretty reliable, and results in less `isTestMode` switching in our
main code, which is great. We do currently still need to run our
coroutines in blocking mode with it, but there may be a solution to
that....

This commit also fixes a double-install problem I discovered that was
causing a lot of problems in test: since `install()` was not idempotent,
we were getting duplicate callbacks and stuff that were the cause of
some unreliability in the tests.